@@ -1,4 +1,9 @@
-export interface Post {
-  title: string;
-  datetime: string;
+export interface IState {
+  user: undefined;
+  userCanvas: string[];
+}
+
+export interface IUserLine {
+  x: number;
+  y: number;
 }
