@@ -1,7 +1,7 @@
 <template>
   <v-app class="home">
     <v-main>
-      <the-draw-main></the-draw-main>
+      <the-draw-main />
     </v-main>
   </v-app>
 </template>
@@ -10,7 +10,6 @@
   import Vue from "vue";
   import TheDrawMain from "@/components/TheDrawMain.vue";
 
-  // import { BACK_URL } from '@/env';
   // import { Post } from '@/interfaces';
 
   export default Vue.extend({
